+++ engine/command_dispatcher.go
@@ -75,8 +75,6 @@ func (d *CommandDispatcher) Execute(cmd string, args []string) (CommandResult, e
 		return d.handleSpec(args)
 	case "--kanban":
 		return d.handleKanban(args)
-	case "backlog":
-		return d.handleBacklog(args)
 	case "commit":
 		return d.handleCommit(args)
 	case "--install-shortcut":